# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCG Building Violation Management System API
  description: REST API consumed by the MCG web portal module and the MCG HARYANA mobile app (building-violations module).
  version: 2.0.0
paths:
...
  /building-violations/api/media/:
    get:
      tags:
      - building-violations
      summary: List Media
      operationId: list_media_building_violations_api_media__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - building-violations
      summary: Upload
      operationId: upload_building_violations_api_media__post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_building_violations_api_media__post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /building-violations/api/media/{pk}/:
    get:
      tags:
      - building-violations
      summary: Media Detail
      operationId: media_detail_building_violations_api_media__pk___get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - building-violations
      summary: Media Delete
      operationId: media_delete_building_violations_api_media__pk___delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /building-violations/api/masters/wards/geojson/:
    get:
      tags:
      - building-violations
      summary: Wards Geojson
      operationId: wards_geojson_building_violations_api_masters_wards_geojson__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /building-violations/api/masters/legal-sections/statutes/:
    get:
      tags:
      - building-violations
      summary: Statutes
      operationId: statutes_building_violations_api_masters_legal_sections_statutes__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    Body_upload_building_violations_api_media__post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        kind:
          type: string
          title: Kind
          default: INSPECTION
        case:
          anyOf:
          - type: string
          - type: 'null'
          title: Case
        notice:
          anyOf:
          - type: string
          - type: 'null'
          title: Notice
        sanctioned_plan:
          anyOf:
          - type: string
          - type: 'null'
          title: Sanctioned Plan
        task:
          anyOf:
          - type: string
          - type: 'null'
          title: Task
        latitude:
          anyOf:
          - type: string
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: string
          - type: 'null'
          title: Longitude
        accuracy_m:
          anyOf:
          - type: string
          - type: 'null'
          title: Accuracy M
        altitude_m:
          anyOf:
          - type: string
          - type: 'null'
          title: Altitude M
        captured_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Captured At
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        caption:
          anyOf:
          - type: string
          - type: 'null'
          title: Caption
        location_integrity:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Integrity
      type: object
      required:
      - file
      title: Body_upload_building_violations_api_media__post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /building-violations/api/auth/otp/verify/
servers:
- url: /
  description: Same origin (the module is mounted at /building-violations/)